requirements.txt
Django==3.1
psycopg2-binary
djangorestframework
Pillow
python-decouple
django-phonenumber-field
phonenumbers